# [v3] Reduce calulate_wake calls in SR yaw optimizer