<!-- Is this pull request ready to be merged? Yes -->
<!-- i.e. tests pass or are expected to fail; all development is finished; appropriate documentation is included. -->
<!-- If not but opening the pull request will facilitate development, make it a "draft" pull request -->

**Feature or improvement description**
This PR replaces a for-loop of floris calls with a single floris call in the yaw optimizer class. This reduces the computation time for `09_optimize_yaw.py` from 20 s to 12 s for free. In a secondary test case, I found that the computation time was only 60% of what it was in the current `v3` implementation (1.7x faster) with the exact same results.

**Related issue, if one exists**
None.

**Impacted areas of the software**
Wake steering optimization.

**Additional supporting information**
-

**Test results, if applicable**
Running `09_optimize_yaw.py` for the current `v3` branch:
```
Optimization finished in 20.01 seconds.
Optimal AEP: 434.878 GWh. Time spent: 4.9 s.
Relative AEP uplift by wake steering: 2.903 %.
```

Running it with this PR:
```
Optimization finished in 11.82 seconds
Optimal AEP: 434.878 GWh. Time spent: 4.4 s.
Relative AEP uplift by wake steering: 2.903 %.
```

I also checked that the results for `08_opt_yaw_single_ws.py` are still identical. For the `v3` branch:
![results_main](https://user-images.githubusercontent.com/22119448/150868192-46cbc736-98fd-4a1b-a032-5c2669871dc4.png)

And with the PR branch:
![results_patch](https://user-images.githubusercontent.com/22119448/150868225-3dae7c93-d8c4-4620-943a-7e7e64c18e06.png)

They are indeed identical.


Finally I ran additional benchmarks with:
```
from time import perf_counter as timerpc
import numpy as np
from floris.tools import FlorisInterface
from floris.tools.optimization.yaw_optimization.yaw_optimizer_sr import YawOptimizationSR


def load_floris(N=5):
    # Load the default example floris object
    fi = FlorisInterface("inputs/gch.yaml") # GCH model matched to the default "legacy_gauss" of V2
    # fi = FlorisInterface("inputs/cc.yaml") # New CumulativeCurl model

    # Specify wind farm layout and update in the floris object
    X, Y = np.meshgrid(
        5.0 * fi.floris.grid.reference_turbine_diameter * np.arange(0, N, 1),
        5.0 * fi.floris.grid.reference_turbine_diameter * np.arange(0, N, 1),
    )
    fi.reinitialize(layout=(X.flatten(), Y.flatten()))

    return fi


if __name__ == "__main__":
    # Load FLORIS
    fi = load_floris(N=7)
    fi.reinitialize(wind_speeds=8.0)
    nturbs = len(fi.layout_x)

    start_time = timerpc()
    fi.reinitialize(
        wind_directions=np.arange(0.0, 360.0, 5.0),
        wind_speeds=[8.0]
    )
    yaw_opt = YawOptimizationSR(
        fi=fi,
        minimum_yaw_angle=0.0,  # Allowable yaw angles lower bound
        maximum_yaw_angle=20.0,  # Allowable yaw angles upper bound
        Ny_passes=[5, 4],
        exclude_downstream_turbines=True,
        exploit_layout_symmetry=True,
    )

    df_opt = yaw_opt._optimize()
    end_time = timerpc()
    t_tot = end_time - start_time
    t_fi = yaw_opt.time_spent_in_floris

    print("Optimization finished in {:.2f} seconds.".format(t_tot))
    ul = df_opt["farm_power_opt"].sum() / df_opt["farm_power_baseline"].sum()
    print("Total farm power uplift: {:.5f} %".format(100.0 * (ul - 1)))
```

And found:
```
nrel/v3 branch:
N=35: 19.42 seconds, 3.76090 %
N=36: 45.17 seconds, 4.25686 %
N=49: 93.39 seconds, 4.66433 %
```
```
PR branch:
N=25: 11.41 seconds, 3.76090 %
N=36: 35.94 seconds, 4.25686 %
N=49: 90.11 seconds, 4.66433 %
```

The benefit is really pronounced at smaller farms (close to 2x speedup), and then converges to pretty much equal performance (~1x, no speedup) as the farm grows past N=49.

<!-- Release checklist:
- Update the version in
    - [ ] README.rst
    - [ ] docs/index.rst
    - [ ] floris/VERSION
- [ ] Verify readthedocs builds correctly
- [ ] Create a tag in the NREL/FLORIS repository
-->